agent/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langgraph
redis
chromadb
pydantic
python-dotenv
